Development
Timestamps should work, but there is a problem with appending logs.

+++ calendar.h
@@ -10,7 +10,7 @@ class Calendar {
         void test_1();
         time_t TimeFromPrevious(time_t & A, time_t & B);
         std::string GetReadableTime ( time_t & A);
-        size_t GetTime();
+        int GetTime();
         std::string GetTime2();
